requirements.txt
django
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0
pytest-django==4.8.0
django-redis==5.4.0

whitenoise  # https://github.com/evansd/whitenoise
redis  # https://github.com/redis/redis-py
hiredis  # https://github.com/redis/hiredis-py
celery  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat  # https://github.com/celery/django-celery-beat
flower  # https://github.com/mher/flower
# Django
# ------------------------------------------------------------------------------
django-environ  # https://github.com/joke2k/django-environ
django-redis  # https://github.com/jazzband/django-redis
# Django REST Framework
django-cors-headers  # https://github.com/adamchainz/django-cors-headers
pyotp  # https://github.com/pyotp/pyotp
pydantic

# Code quality
# ------------------------------------------------------------------------------
flake8  # https://github.com/PyCQA/flake8
flake8-isort  # https://github.com/gforcada/flake8-isort
coverage  # https://github.com/nedbat/coveragepy
black # https://github.com/psf/black
djlint  # https://github.com/Riverside-Healthcare/djLint
pylint-django  # https://github.com/PyCQA/pylint-django
pylint-celery # https://github.com/PyCQA/pylint-celery
pre-commit  # https://github.com/pre-commit/pre-commit


gunicorn  # https://github.com/benoitc/gunicorn

# Monitoring
# -----------------------------------------------------------------------------
sentry-sdk[django]  # sentry-sdk[django]
six